amelioration gestion erreur list user

+++ src/features/backoffice/users/hooks/useDeleteUser.ts
@@ -1,6 +1,7 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { deleteUser } from "../../../../entities/user/api/user.api";
 import { usersQueryKey } from "./useUsers";
+import { AppError } from "../../../../shared/errors/AppError";
 
 export function useDeleteUser(){
     const queryClient = useQueryClient();
@@ -9,6 +10,14 @@ export function useDeleteUser(){
         mutationFn: ( params: {
             userId: number,
         }) => {
+            if (params.userId === -1) {
+                throw new AppError({
+                    message: "Missing userId in delete user mutation",
+                    userMessage: "Utilisateur non trouvé.",
+                    code: "VALIDATION_ERROR",
+                })
+            }
+
             return deleteUser(params.userId);
         },
         onSuccess: () => {
@@ -16,6 +25,5 @@ export function useDeleteUser(){
                 queryKey: usersQueryKey,
             })
         }
-
     })  
 }
